Add FhirAllTypes to the Satellite projects and remove it in Core
